Add support for fetch mocking with `undici` `MockAgent`
Adds `fetchMock` to the `cloudflare:test` module. This is an instance
of `undici`'s `MockAgent`, and can be used to mock global `fetch()`
requests in the test worker. This works by bundling just the `undici`
`MockAgent` code, with a customisable `Dispatcher` that is set to the
original workers `fetch()` function in the pool.

Closes DEVX-1059