Add Durable Object test helpers and support for self-service bindings
Adds a `runInDurableObject()` test helper for running arbitrary code
inside a Durable Object's I/O context, and a `runDurableObjectAlarm()`
helper for running a Durable Object's scheduled alarms.

These require the new `main` option to be set to the worker's
entrypoint, so the Vitest pool can create instances of the user's
Durable Objects.

The `main` option also allows service bindings to be declared to the
worker running the tests. Events dispatched to these service bindings
will use the handler exported by the `main` worker.

Closes DEVX-1056, DEVX-1057 and DEVX-1058